Move module files from util/ to module/
