+++ packages/ag-charts-community/src/chart/background/background.ts
@@ -1,8 +1,8 @@
 import { Rect } from '../../scene/shape/rect';
 import { Group } from '../../scene/group';
-import type { ModuleInstance } from '../../util/baseModule';
-import { BaseModuleInstance } from '../../util/module';
-import type { ModuleContext } from '../../util/moduleContext';
+import type { ModuleInstance } from '../../module/baseModule';
+import { BaseModuleInstance } from '../../module/module';
+import type { ModuleContext } from '../../module/moduleContext';
 import { ProxyPropertyOnWrite } from '../../util/proxy';
 import { BOOLEAN, OPT_COLOR_STRING, Validate } from '../../util/validation';
 import type { LayoutCompleteEvent } from '../layout/layoutService';
