+++ packages/ag-charts-community/src/chart/background/backgroundModule.ts
@@ -1,5 +1,5 @@
 import { Background } from './background';
-import type { Module } from '../../util/module';
+import type { Module } from '../../module/module';
 
 export const BackgroundModule: Module = {
     type: 'root',
