+++ packages/ag-charts-community/src/chart/series/cartesian/scatterSeries.ts
@@ -22,7 +22,7 @@ import type {
     AgTooltipRendererResult,
     AgCartesianSeriesMarkerFormat,
 } from '../../../options/agChartOptions';
-import type { ModuleContext } from '../../../util/moduleContext';
+import type { ModuleContext } from '../../../module/moduleContext';
 import type { DataController } from '../../data/dataController';
 import { createDatumId, diff } from '../../data/processors';
 import { getMarkerConfig, updateMarker } from './markerUtil';
