+++ packages/ag-charts-community/src/chart/series/polar/polarSeries.ts
@@ -4,7 +4,7 @@ import type { BBox } from '../../../scene/bbox';
 import { ChartAxisDirection } from '../../chartAxisDirection';
 import type { PointLabelDatum } from '../../../util/labelPlacement';
 import type { DataModel, ProcessedData } from '../../data/dataModel';
-import type { ModuleContext } from '../../../util/moduleContext';
+import type { ModuleContext } from '../../../module/moduleContext';
 
 export abstract class PolarSeries<S extends SeriesNodeDatum> extends Series<SeriesNodeDataContext<S>> {
     /**
