+++ packages/ag-charts-community/src/module-support.ts
@@ -2,17 +2,17 @@ export * from './util/array';
 export * from './util/validation';
 export * from './util/default';
 export { extractDecoratedProperties, isDecoratedObject, listDecoratedProperties } from './util/decorator';
-export * from './util/baseModule';
-export * from './util/coreModules';
-export * from './util/optionModules';
-export * from './util/module';
-export * from './util/moduleContext';
 export * from './util/number';
 export * from './util/object';
 export * from './util/proxy';
 export * from './util/shapes';
 export * from './util/types';
 export * from './util/theme';
+export * from './module/baseModule';
+export * from './module/coreModules';
+export * from './module/optionModules';
+export * from './module/module';
+export * from './module/moduleContext';
 export * from './chart/background/backgroundModule';
 export * from './chart/chartAxisDirection';
 export { assignJsonApplyConstructedArray } from './chart/chartOptions';
