+++ packages/ag-charts-community/src/module/coreModules.ts
