+++ packages/ag-charts-community/src/module/module.ts
